maitnain server-side ydoc for rendering client page with initial state of the object
